ec2-test-framework enhancements and graviton 4 testing  (#1715)
1. Some of our ec2 instances from the ec2-test-framework were failing to
be properly stopped. This adds logic to the job pruner to stop hanging
instances.
2. Also parallelizing the test runs. GV2 is significantly slower which
causes the run to be much longer. This also makes it much easier for us
to extend testing against new test scripts.
3. Also added graviton 4 testing to the ec2-test-framework.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.