Lower required Go version, add CI test for specific version (#1725)
### Description of changes: 
Add back support for Go 1.17.13 which is used by NetOS. Fix a few
language compatibility issues for interface{} and cmd struct. This
change does not affect the integrity value of the FIPS module.

### Testing:
This change adds a new CI test which uses Go 1.17.13. 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

---------

Co-authored-by: Sean McGrail <mcgrails@amazon.com>

---------

    Revert "Replace interface{} with any"

    This reverts commit 6a90c15f3f901b90acb91fcb4c11cf4c5f3bcfff.