add support for OCSP_SINGLERESP functions (#1703)
Add support for ASN1 and X509v3 extension functions for
`OCSP_SINGLERESP`. These are all consumed in Ruby's OCSP module.

New tests for the extension functions and `OCSP_SINGLERESP` creation.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.