add support for OCSP_copy_nonce (#1711)
Ruby consumes `OCSP_copy_nonce`, which copies the nonce from the request
to the OCSP response.
OpenSSL's `OCSP_copy_nonce` directly appends the new OCSP nonce to the
list of extensions in the response. This allows multiple OCSP nonces to
exist the same response since the existing nonces are not deleted. It's
a bit strange to allow multiple nonces to exist in a single response. To
make things stranger, `OCSP_check_nonce` directly takes the first OCSP
nonce in the list to compare, so any newer nonces are ignored.
We ultimately decided to follow OpenSSL's faulty implementation and
document/test the behavior, there doesn't seem to be much value in
diverging from the original and will only introduce more unanticipated
behavioral changes.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.