Specifying CPU threads in cmake_build.sh to fix CI failures (#1740)
### Description of changes: 
A previous PR (https://github.com/aws/aws-lc/pull/1735/files) broke the
CI but was merged in since the failed checks were not required. This PR
fixes these failures by explicitly setting the number of threads to
parallelize the make command (as opposed to using -j to spin up as many
processes as possible).

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.