Enabling DIT flag in AArch64. (#1687)
- Add a compile flag to enable it.
- DIT APIs and macro are available for external use.
- Add an option `-dit` to the speed test that's available in a build that used the compile flag. 
Enabling it makes the benchmarks closer to not having DIT enabled. 
This is to suggest that the macro can be used by AWS-LC callers in their own scope 
around multiple calls to the library to improve the performance since, 
in that case, DIT would not switch on and off at the entry/exit of the library functions.