Fix for BIO_gets and update documentation (#1756)
### Description of changes: 
BIO_gets takes a len/size parameter which is of type int and can be
negative. Previously, this param was passed into callback_ex without a
check for len being positive. This is an issue since callback_ex defines
this size/len param as a size_t var. Passing in a negative number to
size_t could lead to unexpected behavior.
A similar issue exists in BIO_read and BIO_write

This PR moves the check for len to before calling the callback and
updates documentation for BIO_gets, BIO_read, and BIO_write.