Add word-level popcount operation
The new function word_popcount returns the number of set (equal to 1)
bits in a 64-bit word argument. It's implemented using a somewhat
typical "Hacker's Delight" style combination of integer operations,
not via any specialized instructions. (On later x86 versions, for
example, there is a single POPCNT instruction, but we avoid it to
keep things working more generally.)