Add macros for HMAC precomputed key sizes (#1745)
These macros help users of AWS-LC to know the HMAC precomputed key sizes
without having to create an HMAC context.

Note that we did not define them as `xxx_CHAINING_LENGTH * 2`, where
`xxx` is the hash function name, because `xxx_CHAINING_LENGTH` is an
internal macro we do not want to expose. Instead, we use a static assert
in `hmac.c` to verify the macros are consistent.y.