Allow aarch64 CPUID capability check for all Linux platforms (#1762)
### Issues:
Addresses P147929287

### Description of changes: 
Removes the `OPENSSL_ANDROID` guard around the aarch64 CPUID capability
check and allows it for all aarch64 linux platforms. This is now inline
with the original commit taken from main:
https://github.com/aws/aws-lc/commit/24cb055cb48ae15b9b924ec939f4b993380723f7

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.