Support OCSP_basic_add1_nonce (#1736)
Ruby happens to depend on `OCSP_basic_add1_nonce` which is identical to
`OCSP_request_add1_nonce`, but does the operation on the
`OCSP_BASICRESP` instead. This extends upon the same underlying
function/logic as the identical function.

Add the same test coverage as what we had for `OCSP_request_add1_nonce`,
but with a random nonce with a specified length this time around.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.