Improving instruction flow in `aes_hw_ctr32_encrypt_blocks` tail len = 0 case (#1774)
The change in https://github.com/aws/aws-lc/commit/b929d749f3d09995aa04848cc6fb24f0aec87da4 showed a 0.01 usec increase in EVP-AES-256-CTR on Graviton4. This change optimises the sequence of operations to make use of instruction fusion in the sequence CMP + B.cond.; that is, the `b.le` instruction is placed after the `cmp` instruction, also avoiding executing unneeded instructions if the jump is to be taken.