Add -text support to X509 tool, add Version tool (#1773)
### Description of changes: 
Add the -text option to the X509 tool which prints a certificate's
contents in a human readable format. Also add the version tool which
only supports printing the basic version information.

### Testing:
Added a new test for X509 -text with some minor caveats due to
difference in between OpenSSL versions. I didn't add a test for the
version as that will always be different:
```
AWS-LC:
OpenSSL 1.1.1 (compatible; AWS-LC 1.33.0)

OpenSSL:
OpenSSL 3.3.1 4 Jun 2024 (Library: OpenSSL 3.3.1 4 Jun 2024)
```

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.