Handle systems defining ATOMIC_LONG_LOCK_FREE as expression (#1788)
### Description of changes: 
Fix an issue with some old system that define ATOMIC_LONG_LOCK_FREE as
an expression that uses runtime checks. These are invalid in
preprocessor conditions such as how we swap between the pthread and c11
atomic implementation at compile time.

### Call-outs:
This could be made a runtime check and swap, but that would be a much
larger refactor and that might introduce it's own performance overhead.
Customers with older systems are strongly encouraged to upgrade.

### Testing:
Verified legacy build works, and modern compilers still get the
performance improvement.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.