Reduce collision probability for variable names (#1804)
Using names such as d for global variables works if they are only ever used in small scope. However, Ed25519 is being moved under the fips module, which increase the scope by a lot. To allow this, reduce the likelihood of variable name collisions by prefixing variable names with some context information. In this case k25519, following convention in the rest of the file.