Refactor ENGINE API and memory around METHOD structs (#1776)
### Description of changes: 
1. Breaking changes to the ENGINE API - made the function signatures
match OpenSSL. Removed method_size as a parameter since the method can
no longer be statically allocated.
2. Got rid of old code supporting reference counting for METHOD structs
as it never existed in OpenSSL and was no-ops/unused in BoringSSL
3. Removed static allocation enforcement for METHOD structs. Now
consumers can only dynamically allocate these structs via funcs like
EC_KEY_METHOD_new and RSA_METHOD_new which are added
[here](https://github.com/aws/aws-lc/pull/1785) and
[here](https://github.com/aws/aws-lc/pull/1790).

### Call-outs:
First PR in a series of 3 to refactor and expand support for ENGINE,
RSA_METHOD, and EC_KEY_METHOD structs

### Testing:
No significant tests can be written yet. The memory model will be tested
further in the PRs that follow.
 
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.