Check at runtime that the tool is loading the same libcrypto it was built with (#1716)
### Description of changes: 
The tool CMakeList was already checking /lib and /lib64 for the
different versions of OpenSSH. However, at runtime we set the
LD_LIBRARY_PATH to only /lib which doesn't have the OpenSSL 3+
artifacts. The benchmark tool was then finding the system install of
OpenSSL which happened to be 3.something and close enough to work. This
change is inspired by a similar mechanism OpenSSL's own tool uses to
check build/runtime dependencies.

### Callouts
This uses SSLeay which seems to be the best common way to check at
runtime all the library versions we test against.

### Testing:
Tested locally changing the expected version and observing the tool
fail.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

Co-authored-by: Justin W Smith <103147162+justsmth@users.noreply.github.com>