Move curve25519 implementations to fips module except spake25519 (#1809)
Moves curve25519 backend, x25519, and ed25519 implementations to fips
module. spake25519 is not moved to save on space in `bcm.o`.

This is mostly an exercise in moving files around and fixing up paths.

Note that the fiat code is actually stored in `/third_party/fiat` which
is used for the C-native implementation of the algorithms. But this is
still included in the fips module because the code is included in the
25519-source code files at pre-processing time. So, it's in scope of the
integrity check.