Remove custom PKCS7 ASN1 functions, add new structs (#1726)
### Issues:
Resolves #CryptoAlg-2493

### Description of changes: 

This PR is the first in a series implementing Ruby's supported subset of
the PKCS7 standard. To do this, we remove AWS-LC's customized ASN.1
serialization logic and delegate to `asn1.h` and `asn1t.h`. We also
update the various `PKCS_type_is_*` no-op functions to actually check
the input's type.

The PKCS7 tests currently contain a test signedData structure that is
encoded using BER. Because our `ASN1_get_object` currently
[disallows](https://github.com/aws/aws-lc/blob/8080ce32b118746475c718d14a609bd1325166e1/crypto/asn1/asn1_lib.c#L136)
indefinite-length BER, we need to modify `d2i_PKCS7` to detect and
convert BER (if present) into DER before parsing. This allows us to
retain backwards compatibility with BER-encoded PKCS7 objects. Please
see the PR description on GitHub for more discussion on this topic.

The next PR in this series will implement the various getters, setters,
and allocation functions needed to support CRruby with testing for each.
Please see PR #1780 for an idea of what that will look like.