Update x509 tool to write all output to common BIO which is a file or stdout (#1800)
### Description of changes: 
Previously only some output followed the user's request using `-out` to
direct the output to a file. This change updates everything to use a
common `BIO` which is `stdout` by default or the user provided file.

### Testing:
Added new tests that test writing to a file directly with `-out` we have
existing tests that cover writing to `stdout` (and then getting
redirected to a file with `>`) works as expected.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.