Add KDA OneStep (SSKDF_digest and SSKDF_hmac) to FIPS indicator (#1793)
### Description of changes: 
Adds the FIPS service indicator APIs to `SSKDF_digest` and `SSKDF_hmac`.

**NOTE**: AWS-LC doesn't currently support using SHA3-* digest
algorithms with our HMAC APIs. So that is why it is missing from the
approved digest list for `SSKDF_hmac`.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.