Improve pre-sandbox setup (#1825)
### Description of changes: 

When we added the snapsafe UBE we forgot to setup the pre-jail. Try to
prevent that again by invoking the top-level function `RAND_bytes` which
should execute all the lazy workflows that configures troublesome jail
code.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.