Add OCSP round trip integration test with minor fixes (#1811)
Now that we have support for OCSP responder functions, we can now do a
round trip integration test with an OCSP request from an OCSP client <->
OCSP response from an OCSP responder. There were a couple bugs found
along the way and this implements fixes along with the new tests.
1. `OCSP_cert_to_id` allows for a NULL `subject`, then passes a NULL
`serialNumber` to `OCSP_cert_id_new`. OpenSSL allows the NULL parameter,
but we disallow it. Changed to allow NULL for better interoptability.
2. Our implementation of `X509_gmtime_adj` happens to use UTC Time, but
the producedAt field for OCSP `OCSP_RESPDATA` expects generalized time.
This causes a parsing failure from OCSP responses we generated. This was
pinned down to `X509_gmtime_adj` calling `ASN1_TIME_adj` internally,
which allocates UTCTime if it fits.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.