Move KEM API and ML-KEM definitions to FIPS module (#1828)
In this PR, we first move the ML-KEM definitions under the FIPS module
and the set of KEM API. There are two objects to keep track of:

`KEM` is the "context" object of a kem method storing e.g. size
information, unique identifiers, and a reference to a method table (of
type `KEM_METHOD`).

`KEM_METHOD` is the method table that stores references to
implementations for each abstract operation.

Both `KEM` and `KEM_METHOD` has been defined outside the FIPS module.
Move this under the module now. Previously, also the KEM method function
tables were defined in their own source code file. Move these together
with the definitions of the KEMs. They will later go to their own file.

Note, that `KEM` was defined with only `const` fields. I think this was
my invention but it was not an optimal choice. It only allows static
initialization. This is not good with the current static FIPS build, so
remove the `const` qualifier from all non-pointer fields in `KEM`.

Finally, drop part of the table search method. KEMs are now directly
referenced instead of searching through a table. Retain the table search
for the legacy Kyber kems for now though to minimize code changes in
each PR.