fix socat integration CI (#1833)
socat started cutting off the tests prematurely if the test hadn't
finished yet with these commits:
* 4d03b37da1226f5d76d1503487d7cd8b90c4d80
* b5b9ee0031eb0f38a71b44a45723d728265ee1bd

A lot of tests are related to specific warnings that socat emits based
on the openssl version not supporting certain things. My suspicion is
that the additional time when writing to the logs breaks the 3
millisecond threshold and socat determines that we've failed the test.
Extending this time threshold to 50 milliseconds gets past things.
  
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.