Pre jail unit test (#1835)
#1825 improved the sandbox mode. This PR adds a unit test to catch obvious sandbox configuration oversights.

It works by creating a small application that installs a seccomp filter. Any syscalls, except a small subset, invoked by AWS-LC will trigger the filter. We both check that the filter will trip on AWS-LC without configuration sandbox mode and that the filter doesn't trip when sandbox mode is configured.