More minor symbols Ruby depends on (#1837)
We noticed three missing symbols after testing where we were at with our
support for Ruby. These are all pretty minor and are related to work
we've already done.
* The underlying work for `d2i_ECPKParameters_bio` and
`i2d_ECPKParameters_bio` was already completed with 6056999.
Coverage for actually parsing is included in that PR, so I've only added
brief tests to ensure that the same behavior applies to the BIO methods.
* `RAND_egd_bytes` is for querying the entropy gathering daemon EGD on
the socket path. We already no-op `RAND_egd` (which directly calls
`RAND_egd_bytes` in OpenSSL), so I've applied the same no-op behavior
here as well.

### Testing:
Minor tests for `d2i_ECPKParameters_bio` and `i2d_ECPKParameters_bio`.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.