Add basic P-521 Jacobian coordinate scalar multiplication
The new function p521_jscalarmul[_alt] is the NIST P-521 analog
of the corresponding P-256 and P-384 functions: it does scalar
multiplication of a point on the P-521 curve, where both input and
output points are in Jacobian representation. But in contrast to
the P-256 and P-284 functions, Montgomery form is *not* used.

This update also fixes a few miscellaneous issues in tests and
improves the specs of some other P-521 functions by fixing the
nominal size parameter in "modular_decode". (It is actually not
semantically relevant, being there to maintain the analog with
"montgomery_decode".)