Update Service Indicator to handle custom crypto through *_METHOD structs (#1857)
### Issues:
`CryptoAlg-2633`

### Description of changes: 
The service indicator "update" functions were modified to check for the
existence of custom crypto in RSA_METHOD and EC_KEY_METHOD structs. If
custom crypto is utilized for a given operation, the service indicator
should not be updated. The checks are conservative and do not update the
service indicator for a signing operation for EC/RSA keys if either
custom sign or sign_sig/sign_raw functionality is provided - regardless
of whether these may be invoked.

### Testing:
Tested both EC_KEY_METHOD and RSA_METHOD with custom crypto and ensured
the service indicator was not updating.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.