Fix flaky ssl BadKemKeyShare tests (#1876)
The way the test was generating a public key that is not consistent with
a secret key is by xor-ing the first byte of the key with 1. Such key
modifications can inadvertently make the key invalid and thus fail the
test.

For example, before performing encapsulation ML-KEM decodes the public
key bytes to an array of 12-bit coefficients and checks that all
coefficients are in the range [0, 3328]. If the first two bytes of the
key encode the coefficient 3328 then xor-ing the first byte with 1 will
make the coefficient equal to 3329. The call to encapsulate will then
fail because 3329 is an invalid coefficient.