Remove nginx-tests patch now that upstream supports AWS-LC (#1898)
### Description of changes: 
All of these changes have been applied to upstream nginx-tests in
https://github.com/nginx/nginx-tests/commit/52bf3505da8d5192f0cd6fce331d8c84e1b61f44
and is no longer necessary. The CodeBuild limitation remains so
`nginx_patch_tests` is still called to delete the incompatible test.

This will get the integration tests back to a passing state. 

### Testing:
In this CI run.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.