Fix OCSP timebomb in tests (#1891)
Resolves https://github.com/aws/aws-lc/issues/1889

It turns out the first call to `OCSP_check_validity` in `TestGoodOCSP`
will emit two additional errors on to the error stack once the "this
update" field is outdated. Clearing the previous additional errors on
the stack allows us to check for which specific errors were emitted more
accurately.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.