Remove special s2n-bignum symbol handling sauce from build (#1903)
Symbols from s2n-bignum needed special build shenanigans because they essentially are "third-party" (although still AWS owned) in the context of what code AWS-LC repository can touch. Therefore they didn't have the special "prefix" header files. This corrupts the prefix build in two ways that needed to be handled:

references to s2n-bignum files are prefixed, but the definitions are not. Obviously, this doesn't work...
when we test the prefix build, we must take care to not assume s2n-bignum symbols are prefixed...
Removing these two corruptions, means it's now a bit easier to import s2n-bignum code. A small step closer to maybe switching to a submodule.